@@ -1,20 +1,18 @@
-import React, { Component } from 'react';
-import { Route } from 'react-router';
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
 import { Layout } from './components/Layout';
 import { Home } from './components/Home';
 import { FetchData } from './components/FetchData';
 import { Counter } from './components/Counter';
 
-export default class App extends Component {
-  static displayName = App.name;
-
-  render () {
-    return (
-      <Layout>
-        <Route exact path='/' component={Home} />
-        <Route path='/counter' component={Counter} />
-        <Route path='/fetch-data' component={FetchData} />
-      </Layout>
-    );
-  }
+export default function App() {
+  return (
+    <Layout>
+      <Routes>
+        <Route path='/' element={<Home />} />
+        <Route path='/counter' element={<Counter />} />
+        <Route path='/fetch-data' element={<FetchData />} />
+      </Routes>
+    </Layout>
+  );
 }
