@@ -1,31 +1,21 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 
-export class Counter extends Component {
-  static displayName = Counter.name;
+export function Counter() {
+  const [currentCount, setCurrentCount] = useState(0);
 
-  constructor (props) {
-    super(props);
-    this.state = { currentCount: 0 };
-    this.incrementCounter = this.incrementCounter.bind(this);
-  }
+  const incrementCounter = () => {
+    setCurrentCount(currentCount + 1);
+  };
 
-  incrementCounter () {
-    this.setState({
-      currentCount: this.state.currentCount + 1
-    });
-  }
+  return (
+    <div>
+      <h1>Counter</h1>
 
-  render () {
-    return (
-      <div>
-        <h1>Counter</h1>
+      <p>This is a simple example of a React component.</p>
 
-        <p>This is a simple example of a React component.</p>
+      <p>Current count: <strong>{currentCount}</strong></p>
 
-        <p>Current count: <strong>{this.state.currentCount}</strong></p>
-
-        <button className="btn btn-primary" onClick={this.incrementCounter}>Increment</button>
-      </div>
-    );
-  }
+      <button className="btn btn-primary" onClick={incrementCounter}>Increment</button>
+    </div>
+  );
 }
