@@ -1,20 +1,26 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 
-export class FetchData extends Component {
-  static displayName = FetchData.name;
+export function FetchData() {
+  const [forecasts, setForecasts] = useState([]);
+  const [loading, setLoading] = useState(true);
 
-  constructor (props) {
-    super(props);
-    this.state = { forecasts: [], loading: true };
+  useEffect(() => {
+    async function fetchData() {
+      try {
+        const response = await fetch('api/SampleData/WeatherForecasts');
+        const data = await response.json();
+        setForecasts(data);
+        setLoading(false);
+      } catch (error) {
+        console.error('Error fetching weather data:', error);
+        setLoading(false);
+      }
+    }
+    
+    fetchData();
+  }, []);
 
-    fetch('api/SampleData/WeatherForecasts')
-      .then(response => response.json())
-      .then(data => {
-        this.setState({ forecasts: data, loading: false });
-      });
-  }
-
-  static renderForecastsTable (forecasts) {
+  function renderForecastsTable(forecasts) {
     return (
       <table className='table table-striped'>
         <thead>
@@ -39,17 +45,15 @@ export class FetchData extends Component {
     );
   }
 
-  render () {
-    let contents = this.state.loading
-      ? <p><em>Loading...</em></p>
-      : FetchData.renderForecastsTable(this.state.forecasts);
+  let contents = loading
+    ? <p><em>Loading...</em></p>
+    : renderForecastsTable(forecasts);
 
-    return (
-      <div>
-        <h1>Weather forecast</h1>
-        <p>This component demonstrates fetching data from the server.</p>
-        {contents}
-      </div>
-    );
-  }
+  return (
+    <div>
+      <h1>Weather forecast</h1>
+      <p>This component demonstrates fetching data from the server.</p>
+      {contents}
+    </div>
+  );
 }
